misp-twitter/requirements.txt
aiohttp
async-timeout
Flask
Flask-Migrate
flask-slacksigauth
gunicorn
#pymisp
redis
requests
rq
slackclient
slackeventsapi
tweepy
